pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "z-lib-tool"
version = "0.1.0"
description = "Transparent ZIP file operations for Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/z_lib"]

[dependency-groups]
dev = [
    "pillow>=12.1.1",
    "pytest>=9.0.2",
]
